+++ NEWS.md
@@ -2,6 +2,8 @@
 ---------------------
 * Major revisions to `catmap` function
     * Function now returns output as a list object
+    * Removed redundant and superfluous code
+    * Removed `print.all` argument entirely
 * Make package compatible with CRAN
     * Convert Rd files to Roxygen format
     * Fix WARNING associated with `catmapdata` manual